common: command: Add support for $ auto-completion
This commit is heavily based on a backported commit: 689d2b25d4 by Boris
Brezillon. The change from the commit is making the dollar completion
accept the '$' format instead of '${'.

Original commit message

Add the dollar_complete() function to auto-complete arguments starting
with a '$' and use it in the cmd_auto_complete() path such that all
args starting with a $ can be auto-completed based on the available env
vars.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>
[trini: Fix some linking problems]
Signed-off-by: Tom Rini <trini@konsulko.com>