cli: Add null-terminator to the line buffer in appropriate places
The same buffer is most likely used every time cread_line() is called
when parsing the command line.
In the regular flow, the line buffer is null-terminated right after
eol_num, but it didn't happen in the case of ^C.
This caused a bug in the ^W code:
- The code copied characters up to the null-terminator with puts(), but
the terminator could have been overridden if ^C was used. This caused
junk characters to be displayed on screen.

Make ^C terminate the line correctly, change ^W to use putnstr, and also
pre-terminate the buffer before any parsing as a safety measure.